fixed the some logic in generic service
