+++ src/main/services/features/directionsQuestionnaire/directionQuestionnaireService.ts
@@ -42,14 +42,14 @@ const saveDirectionQuestionnaire = async (claimId: string, value: any, direction
     const claim: any = await getCaseDataFromStore(claimId);
     if (claim.directionQuestionnaire) {
       if (parentPropertyName) {
-        claim.directionQuestionnaire[parentPropertyName][directionQuestionnairePropertyName] = value;
+        claim.directionQuestionnaire[parentPropertyName] = {[directionQuestionnairePropertyName]: value};
       } else {
         claim.directionQuestionnaire[directionQuestionnairePropertyName] = value;
       }
     } else {
       const directionQuestionnaire: any = new DirectionQuestionnaire();
       if (parentPropertyName) {
-        directionQuestionnaire[parentPropertyName][directionQuestionnairePropertyName] = value;
+        directionQuestionnaire[parentPropertyName] = {[directionQuestionnairePropertyName]: value};
       } else {
         directionQuestionnaire[directionQuestionnairePropertyName] = value;
       }
